feature: add errors into prometheus /metrics scrape
