+++ server/src/middleware/mod.rs
@@ -1,3 +1,5 @@
 pub mod api_version;
 pub mod auth_middleware;
 pub mod json_middleware;
+pub mod metrics_middleware;
+pub mod timeout_middleware;
